requirements.txt
Flask==1.1.2
numpy==1.18.5
opencv-python==4.4.0.42
Pillow==7.2.0
requests==2.25.0
